+++ builders/storage/StorageS3Builder.d.ts
@@ -1,22 +1,16 @@
-import type { Stats } from "fs";
-import type { StorageDisk, StorageOptions } from "../types/storage";
-export default class StorageBuilder {
-    protected conf: Record<string, any>;
-    protected overrideDisk?: StorageDisk;
-    protected drive?: string;
-    constructor();
+import type { StorageDriver, StorageOptions } from "../../types/storage";
+export default class StorageS3Builder implements StorageDriver {
+    protected _config: Record<string, any>;
+    protected client: Bun.S3Client;
+    constructor(config: Record<string, any>);
     private get config();
-    private get currentDisk();
-    private get driver();
-    build(overrideDisk: StorageDisk): StorageBuilder;
-    disk(drive: string): StorageBuilder;
     exists(filepath: string): Promise<boolean>;
     missing(filepath: string): Promise<boolean>;
-    metadata(filepath: string): Promise<Stats | Bun.S3Stats>;
+    metadata(filepath: string): Promise<Bun.S3Stats>;
     size(filepath: string): Promise<number>;
     mimeType(filepath: string): Promise<string>;
     lastModified(filepath: string): Promise<Date>;
-    get(filepath: string): Promise<Bun.BunFile | Bun.S3File>;
+    get(filepath: string): Promise<Bun.S3File>;
     put(filepath: string, content: any, options?: StorageOptions): Promise<void>;
     copy(source: string, destination: string, options?: StorageOptions): Promise<void>;
     move(source: string, destination: string, options?: StorageOptions): Promise<void>;