+++ facades/Storage.js
@@ -12,11 +12,29 @@ export default class Storage {
     static async missing(path) {
         return await new StorageBuilder().missing(path);
     }
+    static async metadata(path) {
+        return await new StorageBuilder().metadata(path);
+    }
+    static async size(path) {
+        return await new StorageBuilder().size(path);
+    }
+    static async mimeType(path) {
+        return await new StorageBuilder().mimeType(path);
+    }
+    static async lastModified(path) {
+        return await new StorageBuilder().lastModified(path);
+    }
     static async get(path) {
         return await new StorageBuilder().get(path);
     }
     static async put(path, content, options) {
         return await new StorageBuilder().put(path, content);
+    }
+    static async copy(source, destination, options) {
+        return await new StorageBuilder().copy(source, destination, options);
+    }
+    static async move(source, destination, options) {
+        return await new StorageBuilder().move(source, destination, options);
     }
     static async delete(path) {
         return await new StorageBuilder().delete(path);