+++ src/stubs/controllers/TemplateController.ts
@@ -2,22 +2,32 @@ import BaseController from "@bejibun/core/bases/BaseController";
 
 export default class TemplateController extends BaseController {
     public async index(request: Bun.BunRequest): Promise<Response> {
+        // Your code goes here
+
         return super.response.setData().send();
     }
 
     public async store(request: Bun.BunRequest): Promise<Response> {
+        // Your code goes here
+
         return super.response.setData().send();
     }
 
     public async show(request: Bun.BunRequest): Promise<Response> {
+        // Your code goes here
+
         return super.response.setData().send();
     }
 
     public async update(request: Bun.BunRequest): Promise<Response> {
+        // Your code goes here
+
         return super.response.setData().send();
     }
 
     public async destroy(request: Bun.BunRequest): Promise<Response> {
+        // Your code goes here
+
         return super.response.setData().send();
     }
 }